requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
Werkzeug==3.0.2
zipp==3.18.1
pymysql==1.1.0
boto3==1.34.81
requests-toolbelt==1.0.0
pillow==10.3.0


